requirements.txt
python-dotenv
discord.py
PyMuPDF
